Change: Split PortList and PortLists command into two modules
Separate both commands into two modules and add tests for the
PortListsCommand.
