+++ src/gmp/gmp.ts
@@ -57,7 +57,8 @@ import PermissionCommand from 'gmp/commands/permission';
 import PermissionsCommand from 'gmp/commands/permissions';
 import PoliciesCommand from 'gmp/commands/policies';
 import PolicyCommand from 'gmp/commands/policy';
-import {PortListCommand, PortListsCommand} from 'gmp/commands/port-lists';
+import PortListCommand from 'gmp/commands/port-list';
+import PortListsCommand from 'gmp/commands/port-lists';
 import ReportCommand from 'gmp/commands/report';
 import ReportApplicationsCommand from 'gmp/commands/report-applications';
 import ReportClosedCvesCommand from 'gmp/commands/report-closed-cves';
